requirements.txt
Flask==3.0.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
mysqlclient==2.2.4
python-dotenv==1.0.1
